requirements.txt
asn1crypto==0.24.0
catboost==0.4.1
certifi==2018.10.15
cffi==1.11.5
chardet==3.0.4
cryptography==2.3.1
emoji==0.5.1
enum34==1.1.6
future==0.16.0
idna==2.7
numpy==1.15.2
pandas==0.23.4
Pillow==5.3.0
pycparser==2.19
PySocks==1.6.8
python-dateutil==2.7.3
python-telegram-bot==11.1.0
pytz==2018.5
requests==2.20.0
six==1.11.0
staticmap==0.5.4
urllib3==1.24
weather-api==1.0.6
